Fix hang on error when using uv on Windows
